fixup into layers
