+++ rotary/torch-ext/rotary/__init__.py
@@ -1,5 +1,19 @@
-from .ext import apply_rotary_transformers
-from .layers import apply_rotary
+import torch
+
+from ._ops import ops
+from .layers import apply_rotary_transformers
+
+
+def apply_rotary(
+    x1: torch.Tensor,
+    x2: torch.Tensor,
+    cos: torch.Tensor,
+    sin: torch.Tensor,
+    out1: torch.Tensor,
+    out2: torch.Tensor,
+    conj: bool,
+) -> None:
+    ops.apply_rotary(x1, x2, cos, sin, out1, out2, conj)
 
 
 __all__ = ["apply_rotary", "apply_rotary_transformers"]
