--- rotary/torch-ext/rotary/ext/__init__.py
@@ -1,4 +1,0 @@
-from .transformers import apply_rotary_transformers
-
-
-__all__ = ["apply_rotary_transformers"]
